[SVLS-5262] move private Span Pointers functionality to its own file
